Automatic Roles and Unit Test
Adds automatic creation of the client and developer roles to the database to make it a bit more fool proof and make it easier when updating. Also added my own unit test that tests our pages to see if they are getting a good response back and can also simulate an authorized user.